requirements.txt
phidata
streamlit
fastapi
uvicorn
psycopg2-binary